Add model for proposal and change model of user.
