+++ src/main/java/com/intive/patronage/toz/proposals/ProposalController.java
@@ -1,0 +1,4 @@
+package com.intive.patronage.toz.proposals;
+
+public class ProposalController {
+}
