+++ src/main/java/com/intive/patronage/toz/proposals/model/Proposal.java
@@ -1,0 +1,25 @@
+package com.intive.patronage.toz.proposals.model;
+
+import com.intive.patronage.toz.base.model.Identifiable;
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+import java.util.Date;
+
+@Getter
+@Setter
+@Entity
+public class Proposal extends Identifiable{
+
+    @CreationTimestamp
+    @Column(updatable = false)
+    @Temporal(TemporalType.TIMESTAMP)
+    private Date creationDate;
+
+    boolean isRead;
+}
