+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -261,3 +261,36 @@ databaseChangeLog:
                     type: varchar(255)
                     constraints:
                       - nullable: false
+  - changeSet:
+        id : 5
+        author: macieg-b
+        changes:
+          - createTable:
+              tableName: proposal
+              columns:
+                - column:
+                    name: id
+                    type: uuid
+                    constraints:
+                      - primaryKey: true
+                      - nullable: false
+                - column:
+                    name: name
+                    type: varchar(255)
+                - column:
+                    name: surname
+                    type: varchar(255)
+                - column:
+                    name: phone_number
+                    type: varchar(255)
+                - column:
+                    name: email
+                    type: varchar(255)
+                    constraints:
+                      - unique: true
+                - column:
+                    name: creation_date
+                    type: datetime
+                - column:
+                    name: is_read
+                    type: boolean
